+++ src/main/java/org/alfime/apoyo/domain/Rol.java
@@ -7,7 +7,7 @@ import java.util.Set;
 public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Short id;
 
     @Column(length = 50)
     private String nombre;
@@ -15,11 +15,11 @@ public class Rol {
     @ManyToMany(mappedBy = "roles")
     private Set<Asociado> asociados;
 
-    public Long getId() {
+    public Short getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Short id) {
         this.id = id;
     }
